added a room creating page

+++ src/main/java/com/piinfo/db/data/Furniture_SettingsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Furniture_SettingsRepository extends JpaRepository<Furniture_Settings, Integer> {
+public interface Furniture_SettingsRepository extends JpaRepository<Furniture_Settings, Long> {
 	Furniture_Settings findByType (String Type);
 }
